pyproject.toml
[project]
name = "seiaitask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.12.0",
    "numpy>=1.26.0,<2.0.0",
    "scikit-learn>=1.4.0",
    "sentence-transformers>=3.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.0.0",
    "livekit-api>=0.8.0",
    "livekit-agents[deepgram,noise-cancellation,openai,silero,turn-detector]>=1.4.0",
    "livekit-plugins-piper-tts>=0.1.0",
    "twilio>=9.10.3",
    "livekit-plugins-noise-cancellation>=0.2.5",
]
